requirements.txt
oracledb==2.5.1
python-dotenv==1.0.1

